// libs/features/collection/billing/src/lib/shared/models/mof-employeShare-summary.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class MofEmployeShareSummary {
  totalContribution: number = undefined;
  totalDebitAdjustment: number = undefined;
  totalReceiptsAndCredits: number = undefined;
  RequiredPaymentByMOF: number = undefined;
}